LayoutTests/platform/mac-leopard/fast/text/find-kana-expected.txt

Tests find for strings with kana letters in them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Exact matches first as a baseline

FAIL canFind(decomposedHalfwidthKatakanaLetterBa, decomposedHalfwidthKatakanaLetterBa) should be true. Was false.
FAIL canFind(decomposedHalfwidthKatakanaLetterPa, decomposedHalfwidthKatakanaLetterPa) should be true. Was false.
FAIL canFind(decomposedHiraganaLetterBa, decomposedHiraganaLetterBa) should be true. Was false.
FAIL canFind(decomposedHiraganaLetterGa, decomposedHiraganaLetterGa) should be true. Was false.
FAIL canFind(decomposedHiraganaLetterPa, decomposedHiraganaLetterPa) should be true. Was false.
FAIL canFind(decomposedKatakanaLetterGa, decomposedKatakanaLetterGa) should be true. Was false.
PASS canFind(decomposedLatinCapitalLetterAWithGrave, decomposedLatinCapitalLetterAWithGrave) is true
PASS canFind(halfwidthKatakanaLetterA, halfwidthKatakanaLetterA) is true
PASS canFind(halfwidthKatakanaLetterHa, halfwidthKatakanaLetterHa) is true
PASS canFind(halfwidthKatakanaLetterKa, halfwidthKatakanaLetterKa) is true
PASS canFind(halfwidthKatakanaLetterSmallA, halfwidthKatakanaLetterSmallA) is true
PASS canFind(hiraganaLetterA, hiraganaLetterA) is true
PASS canFind(hiraganaLetterBa, hiraganaLetterBa) is true
PASS canFind(hiraganaLetterGa, hiraganaLetterGa) is true
PASS canFind(hiraganaLetterHa, hiraganaLetterHa) is true
PASS canFind(hiraganaLetterKa, hiraganaLetterKa) is true
PASS canFind(hiraganaLetterPa, hiraganaLetterPa) is true
PASS canFind(hiraganaLetterSmallA, hiraganaLetterSmallA) is true
PASS canFind(katakanaLetterA, katakanaLetterA) is true
PASS canFind(katakanaLetterGa, katakanaLetterGa) is true
PASS canFind(katakanaLetterKa, katakanaLetterKa) is true
PASS canFind(katakanaLetterSmallA, katakanaLetterSmallA) is true
PASS canFind(latinCapitalLetterAWithGrave, latinCapitalLetterAWithGrave) is true

Hiragana, katakana, and half width katakana: Must be treated as equal

FAIL canFind(decomposedHiraganaLetterGa, decomposedKatakanaLetterGa) should be true. Was false.
FAIL canFind(decomposedKatakanaLetterGa, decomposedHiraganaLetterGa) should be true. Was false.
PASS canFind(hiraganaLetterA, halfwidthKatakanaLetterA) is true
PASS canFind(hiraganaLetterA, katakanaLetterA) is true
PASS canFind(katakanaLetterSmallA, hiraganaLetterSmallA) is true

Composed and decomposed forms: Must be treated as equal

PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterBa) is true
FAIL canFind(decomposedHiraganaLetterGa, decomposedKatakanaLetterGa) should be true. Was false.
PASS canFind(decomposedHiraganaLetterGa, hiraganaLetterGa) is true
PASS canFind(decomposedHiraganaLetterGa, katakanaLetterGa) is true
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterPa) is true
FAIL canFind(decomposedKatakanaLetterGa, decomposedHiraganaLetterGa) should be true. Was false.
PASS canFind(decomposedLatinCapitalLetterAWithGrave, latinCapitalLetterAWithGrave) is true
FAIL canFind(hiraganaLetterBa, decomposedHiraganaLetterBa) should be true. Was false.
FAIL canFind(hiraganaLetterGa, decomposedHiraganaLetterGa) should be true. Was false.
FAIL canFind(hiraganaLetterPa, decomposedHiraganaLetterPa) should be true. Was false.
FAIL canFind(katakanaLetterGa, decomposedHiraganaLetterGa) should be true. Was false.
PASS canFind(latinCapitalLetterAWithGrave, decomposedLatinCapitalLetterAWithGrave) is true

Small and non-small kana letters: Must *not* be treated as equal

PASS canFind(halfwidthKatakanaLetterA, hiraganaLetterSmallA) is false
PASS canFind(halfwidthKatakanaLetterSmallA, halfwidthKatakanaLetterA) is false
PASS canFind(hiraganaLetterA, hiraganaLetterSmallA) is false
PASS canFind(hiraganaLetterSmallA, katakanaLetterA) is false
PASS canFind(katakanaLetterA, halfwidthKatakanaLetterSmallA) is false
PASS canFind(katakanaLetterSmallA, katakanaLetterA) is false

Kana letters where the only difference is in voiced sound marks: Must *not* be treated as equal

PASS canFind(decomposedHalfwidthKatakanaLetterBa, halfwidthKatakanaLetterHa) is false
PASS canFind(decomposedHalfwidthKatakanaLetterPa, halfwidthKatakanaLetterHa) is false
PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterHa) is false
PASS canFind(decomposedHiraganaLetterBa, hiraganaLetterPa) is false
PASS canFind(decomposedHiraganaLetterGa, halfwidthKatakanaLetterKa) is false
PASS canFind(decomposedHiraganaLetterGa, hiraganaLetterKa) is false
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterBa) is false
PASS canFind(decomposedHiraganaLetterPa, hiraganaLetterHa) is false
PASS canFind(decomposedKatakanaLetterGa, halfwidthKatakanaLetterKa) is false
PASS canFind(decomposedKatakanaLetterGa, hiraganaLetterKa) is false
FAIL canFind(halfwidthKatakanaLetterHa, decomposedHalfwidthKatakanaLetterBa) should be false. Was true.
FAIL canFind(halfwidthKatakanaLetterHa, decomposedHalfwidthKatakanaLetterPa) should be false. Was true.
FAIL canFind(halfwidthKatakanaLetterKa, decomposedHiraganaLetterGa) should be false. Was true.
FAIL canFind(halfwidthKatakanaLetterKa, decomposedKatakanaLetterGa) should be false. Was true.
PASS canFind(hiraganaLetterBa, decomposedHiraganaLetterPa) is false
PASS canFind(hiraganaLetterBa, hiraganaLetterHa) is false
PASS canFind(hiraganaLetterBa, hiraganaLetterPa) is false
PASS canFind(hiraganaLetterGa, hiraganaLetterKa) is false
FAIL canFind(hiraganaLetterHa, decomposedHiraganaLetterBa) should be false. Was true.
FAIL canFind(hiraganaLetterHa, decomposedHiraganaLetterPa) should be false. Was true.
PASS canFind(hiraganaLetterHa, hiraganaLetterBa) is false
PASS canFind(hiraganaLetterHa, hiraganaLetterPa) is false
FAIL canFind(hiraganaLetterKa, decomposedHiraganaLetterGa) should be false. Was true.
FAIL canFind(hiraganaLetterKa, decomposedKatakanaLetterGa) should be false. Was true.
PASS canFind(hiraganaLetterKa, hiraganaLetterGa) is false
PASS canFind(hiraganaLetterPa, decomposedHiraganaLetterBa) is false
PASS canFind(hiraganaLetterPa, hiraganaLetterBa) is false
PASS canFind(hiraganaLetterPa, hiraganaLetterHa) is false

Composed/decomposed form differences before kana characters must have no effect

PASS canFind(decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterA, latinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA, latinCapitalLetterAWithGrave + halfwidthKatakanaLetterA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterA, latinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa, latinCapitalLetterAWithGrave + hiraganaLetterGa) is true
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa, latinCapitalLetterAWithGrave + hiraganaLetterKa) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterKa, latinCapitalLetterAWithGrave + hiraganaLetterGa) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA, latinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + katakanaLetterA, latinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA) is false
PASS canFind(decomposedLatinCapitalLetterAWithGrave + katakanaLetterSmallA, latinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + halfwidthKatakanaLetterA, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterA, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterGa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa) is true
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterGa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterKa) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterKa, decomposedLatinCapitalLetterAWithGrave + hiraganaLetterGa) is false
PASS canFind(latinCapitalLetterAWithGrave + hiraganaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + katakanaLetterA) is false
PASS canFind(latinCapitalLetterAWithGrave + katakanaLetterA, decomposedLatinCapitalLetterAWithGrave + halfwidthKatakanaLetterSmallA) is false
PASS canFind(latinCapitalLetterAWithGrave + katakanaLetterSmallA, decomposedLatinCapitalLetterAWithGrave + katakanaLetterA) is false

PASS successfullyParsed is true

TEST COMPLETE
